Add FYI messaging on pasting cURL command
Something that gave me a hiccup was that when I clicked a button to copy the cURL command and pasted it to my IDE, I did not see the complete code of cURL command. This is because when I launch my IDE and open terminal, the terminal window is by default small, so it doesn't output the entire code if it is long.

I can see the full pasted code if I either drag the terminal window to become larger, or if I use the arrow up button on the keyboard to "scroll" up to see the beginning of the code. An FYI message like below is helpful:

"If your terminal doesn't show the entire code you copied from docs, either enlarge your terminal window or use arrow-up key on the keyboard to scroll to the beginning of the code"

Attaching the screenshot in the comment to show what I mean.