Update add-data-transformation.mdx
In the current code snippet with email, unintentionally the code is making email address a required field - a feature I think we are not looking to introduce in this code snippet. I modified this to only show errors on invalid values in email address. Otherwise, if the value is just null and first name value is present, it doesn't throw an error.

If you agree with my logic to modify the code to NOT require email field if a value in another field is present, please test on your end to ensure my code executes correctly for you too