Update deploying.mdx
Corrected an issue with the deploy command. It was missing the `@latest` after `npx flatfile`, and it was also missing a `=` after token. I also edited the command in the CLI code snippet to show that the example is consistent with the command we advise to use.

The correct command is found [in this documentation](https://flatfile.com/docs/quickstart/deploying)