Update field-types.mdx
A customer discovered today that if in the `enum` fields they do not provide "label", then even if they map "value" in matching the grid will show "undefined". 

Label must always be passed, and this code snippet suggest this can be omitted. Adding this to code snippet to make it more explicit