Update guide.mdx (#688)
Suggesting a slight edit to inline comments of `--color-pigeon-200:`. The documentation currently says `secondary button color`. Actually, it is a color of the secondary button ONLY WHEN you hover over it, it is not a standing color that is always there. Therefore, updating the description. Suggesting this edit based on me testing on a pop up that appears when trying to close the importer on React.