Update guide.mdx (#687)
Proposing a slight change to the comment. I am playing around with styling, and `--color-pigeon-300:` does not change secondary button color on hover. What it changes is the border around the secondary button, and not on hover - border color stays there (when testing on the confirmation pop up when trying to close an importer).

Co-authored-by: Ashley Mulligan <ashleygmulligan@gmail.com>