Update v3_upgrade.mdx
Removed the "pinned" label (please fight me on this if it should stay in!) as we don't have a similar concept in v3 and so it doesn't make sense as a comparison here. 

Added a note about how the sheet is shown inside a workbook, as while workbooks existed in the PSDK we're only showing an example of a sheet for that syntax.