Update meet-the-listener.mdx (#1023)
Cleaned up some language, and also removed the following line:

```
Remember the secrets key you used to create your first Workbook? Your listener is going to use that same credential to receive Events.

```

In this new tutorial, we are not creating a workbook via cURL. I assume this text is referring to a step from the old tutorial. There is nothing proceeding this page when it comes to creating a workbook, so this sentence seems to be no longer relevant.

Co-authored-by: Ashley Mulligan <ashleygmulligan@gmail.com>